+++ commands/registry.go
@@ -32,5 +32,6 @@ func (r *Registry) RegisterCommands() {
 	rootCmd.AddCommand(r.Listeners().Command)
 	rootCmd.AddCommand(r.Clusters().Command)
 	rootCmd.AddCommand(r.Certificates().Command)
+	rootCmd.AddCommand(r.Config().Command)
 	rootCmd.AddCommand(r.Version().Command)
 }